Added version 2.5.1525

+++ chromium/services/network/http_cache_data_counter.cc
@@ -6,6 +6,7 @@
 
 #include <utility>
 
+#include "base/bind.h"
 #include "base/callback.h"
 #include "base/location.h"
 #include "base/threading/sequenced_task_runner_handle.h"
@@ -78,7 +79,7 @@ void HttpCacheDataCounter::GotBackend(
     return;
   }
 
-  int rv;
+  int64_t rv;
   disk_cache::Backend* cache = *backend;
 
   // Handle this here since some backends would DCHECK on this.
@@ -106,7 +107,7 @@ void HttpCacheDataCounter::GotBackend(
 }
 
 void HttpCacheDataCounter::PostResult(bool is_upper_limit,
-                                      int result_or_error) {
+                                      int64_t result_or_error) {
   base::SequencedTaskRunnerHandle::Get()->PostTask(
       FROM_HERE, base::BindOnce(std::move(callback_), this, is_upper_limit,
                                 result_or_error));
